update settings handling
