+++ meteor/client/layout.js
@@ -49,4 +49,40 @@ if (Meteor.isClient) {
             $('ul:first',$(e.target)).css('visibility', 'visible');
         }
     });
+
+    Template.layout.rendered=function(){
+    // Intercepts all XHRs and reload the main browser window on redirect or request error (such as CORS denying access)
+        // This is because, if you run MozDef behind an access-proxy, the requests maybe 302'd to an authentication
+        // provider, but Meteor does not know or handle this. Reloading the main browser window will send the user to the
+        // authentication provider correctly and follow the 302.
+        // Note that since they're 302's they will ALWAYS cause a CORS error, which we keep as this is the SAFE way to
+        // handle this situation.
+        (function(xhr) {
+            var authenticationType = getSetting('authenticationType').toLowerCase();
+            function intercept_xhr(xhrInstance) {
+                // Verify a user is actually logged in and Meteor is running
+                if ((Meteor.user() !== null) && (Meteor.status().connected)) {
+                    // Status 0 means the request failed (CORS denies access)
+                    if (xhrInstance.readyState == 4 && (xhrInstance.status == 302 || xhrInstance.status == 0)) {
+                            location.reload();
+                    }
+                }
+            }
+            var send = xhr.send;
+            xhr.send = function(data) {
+                var origFunc = this.onreadystatechange;
+                if (origFunc) {
+                    this.onreadystatechange = function() {
+                        // We only start hooking for oidc authentication, as this is the only method that is currently
+                        // REQUIRING an access proxy and thus likely to run into 302s
+                        if (authenticationType == 'oidc'){
+                            intercept_xhr(this);
+                        }
+                        return origFunc.apply(this, arguments);
+                    };
+                }
+                return send.apply(this, arguments);
+            };
+        })(XMLHttpRequest.prototype);
+    }
 }
