+++ meteor/client/main.js
@@ -5,8 +5,8 @@ import { Mongo } from 'meteor/mongo';
 import { Session } from 'meteor/session';
 import { _ } from 'meteor/underscore';
 import { Blaze } from 'meteor/blaze';
-import '/imports/collections.js';
 import '/imports/settings.js';
+import '/imports/collections.js';
 import '/imports/helpers.js';
 import '/imports/models.js';
 import '/client/about.html';
