+++ meteor/imports/collections.js
@@ -9,13 +9,13 @@ import uuid from "uuid";
 
 //collections shared by client/server
 Meteor.startup(() => {
+    mozdefsettings = new Meteor.Collection("mozdefsettings");
     events = new Meteor.Collection("events");
     alerts = new Meteor.Collection("alerts");
     investigations = new Meteor.Collection("investigations");
     incidents = new Meteor.Collection("incidents");
     veris = new Meteor.Collection("veris");
     kibanadashboards = new Meteor.Collection("kibanadashboards");
-    mozdefsettings = new Meteor.Collection("mozdefsettings");
     healthfrontend = new Meteor.Collection("healthfrontend");
     sqsstats = new Meteor.Collection("sqsstats");
     healthescluster = new Meteor.Collection("healthescluster");
@@ -376,9 +376,10 @@ Meteor.startup(() => {
         options={
             _suppressSameNameError : true
         };
+        Meteor.subscribe("mozdefsettings");
         alertsCount = new Meteor.Collection("alerts-count",options);
         //client-side subscriptions to low volume collections
-        Meteor.subscribe("mozdefsettings");
+
         Meteor.subscribe("veris");
         Meteor.subscribe("kibanadashboards");
         Meteor.subscribe("userActivity");
