+++ meteor/server/main.js
@@ -1,8 +1,9 @@
 import { Meteor } from 'meteor/meteor';
 import '/imports/settings.js';
 import '/imports/collections.js'
-import './methods.js';
 import './mozdef.js';
+import './methods.js';
+
 
 Meteor.startup(() => {
   // placeholder for any code to run on server at startup
