+++ meteor/server/mozdef.js
@@ -46,7 +46,7 @@ if (Meteor.isServer) {
         mozdefsettings.insert({
             key: 'authenticationType',
             value: mozdef.authenticationType
-        })
+        });
 
         //allow local account creation?
         //http://docs.meteor.com/#/full/accounts_config
